Rename JsonHelpers to FancyZonesData Add new JsonHelpers.[h|cpp] files
