+++ src/modules/fancyzones/dll/dllmain.cpp
@@ -153,7 +153,7 @@ public:
     {
         app_name = GET_RESOURCE_STRING(IDS_FANCYZONES);
         m_settings = MakeFancyZonesSettings(reinterpret_cast<HINSTANCE>(&__ImageBase), FancyZonesModule::get_name());
-        JSONHelpers::FancyZonesDataInstance().LoadFancyZonesData();
+        FancyZonesDataNS::FancyZonesDataInstance().LoadFancyZonesData();
         s_instance = this;
     }
