+++ src/modules/fancyzones/lib/trace.cpp
@@ -2,7 +2,7 @@
 #include "trace.h"
 #include "lib/ZoneSet.h"
 #include "lib/Settings.h"
-#include "lib/JsonHelpers.h"
+#include "lib/FancyZonesData.h"
 
 TRACELOGGING_DEFINE_PROVIDER(
     g_hProvider,
@@ -71,7 +71,7 @@ void Trace::FancyZones::OnKeyDown(DWORD vkCode, bool win, bool control, bool inM
 
 void Trace::FancyZones::DataChanged() noexcept
 {
-    const JSONHelpers::FancyZonesData& data = JSONHelpers::FancyZonesDataInstance();
+    const FancyZonesDataNS::FancyZonesData& data = FancyZonesDataNS::FancyZonesDataInstance();
     int appsHistorySize = static_cast<int>(data.GetAppZoneHistoryMap().size());
     const auto& customZones = data.GetCustomZoneSetsMap();
     const auto& devices = data.GetDeviceInfoMap();
@@ -82,15 +82,15 @@ void Trace::FancyZones::DataChanged() noexcept
         return;
     }
 
-    auto getCustomZoneCount = [&data](const std::variant<JSONHelpers::CanvasLayoutInfo, JSONHelpers::GridLayoutInfo>& layoutInfo) -> int {
-        if (std::holds_alternative<JSONHelpers::GridLayoutInfo>(layoutInfo))
+    auto getCustomZoneCount = [&data](const std::variant<FancyZonesDataNS::CanvasLayoutInfo, FancyZonesDataNS::GridLayoutInfo>& layoutInfo) -> int {
+        if (std::holds_alternative<FancyZonesDataNS::GridLayoutInfo>(layoutInfo))
         {
-            const auto& info = std::get<JSONHelpers::GridLayoutInfo>(layoutInfo);
+            const auto& info = std::get<FancyZonesDataNS::GridLayoutInfo>(layoutInfo);
             return (info.rows() * info.columns());
         }
-        else if (std::holds_alternative<JSONHelpers::CanvasLayoutInfo>(layoutInfo))
+        else if (std::holds_alternative<FancyZonesDataNS::CanvasLayoutInfo>(layoutInfo))
         {
-            const auto& info = std::get<JSONHelpers::CanvasLayoutInfo>(layoutInfo);
+            const auto& info = std::get<FancyZonesDataNS::CanvasLayoutInfo>(layoutInfo);
             return static_cast<int>(info.zones.size());
         }
         return 0;
@@ -108,15 +108,15 @@ void Trace::FancyZones::DataChanged() noexcept
     std::wstring activeZoneSetInfo;
     for (const auto& [id, device] : devices)
     {
-        const JSONHelpers::ZoneSetLayoutType type = device.activeZoneSet.type;
+        const FancyZonesDataNS::ZoneSetLayoutType type = device.activeZoneSet.type;
         if (!activeZoneSetInfo.empty())
         {
             activeZoneSetInfo += L"; ";
         }
-        activeZoneSetInfo += L"type: " + JSONHelpers::TypeToString(type);
+        activeZoneSetInfo += L"type: " + FancyZonesDataNS::TypeToString(type);
 
         int zoneCount = -1;
-        if (type == JSONHelpers::ZoneSetLayoutType::Custom)
+        if (type == FancyZonesDataNS::ZoneSetLayoutType::Custom)
         {
             const auto& activeCustomZone = customZones.find(device.activeZoneSet.uuid);
             if (activeCustomZone != customZones.end())
