+++ src/modules/fancyzones/tests/UnitTests/JsonHelpers.Tests.cpp
@@ -2,12 +2,12 @@
 #include <filesystem>
 #include <fstream>
 
-#include <lib/JsonHelpers.h>
+#include <lib/FancyZonesData.h>
 #include "util.h"
 
 #include <CppUnitTestLogger.h>
 
-using namespace JSONHelpers;
+using namespace FancyZonesDataNS;
 using namespace Microsoft::VisualStudio::CppUnitTestFramework;
 
 namespace FancyZonesUnitTests
@@ -156,7 +156,7 @@ namespace FancyZonesUnitTests
 
             for (const auto& expected : expectedMap)
             {
-                auto actual = JSONHelpers::TypeToString(static_cast<ZoneSetLayoutType>(expected.first));
+                auto actual = FancyZonesDataNS::TypeToString(static_cast<ZoneSetLayoutType>(expected.first));
                 Assert::AreEqual(expected.second, actual);
             }
         }
@@ -174,7 +174,7 @@ namespace FancyZonesUnitTests
 
             for (const auto& expected : expectedMap)
             {
-                auto actual = JSONHelpers::TypeFromString(expected.second);
+                auto actual = FancyZonesDataNS::TypeFromString(expected.second);
                 Assert::AreEqual(static_cast<int>(expected.first), static_cast<int>(actual));
             }
         }
@@ -195,7 +195,7 @@ namespace FancyZonesUnitTests
 
             for (const auto& expected : expectedMap)
             {
-                auto actual = JSONHelpers::TypeFromLayoutId(expected.second);
+                auto actual = FancyZonesDataNS::TypeFromLayoutId(expected.second);
                 Assert::AreEqual(static_cast<int>(expected.first), static_cast<int>(actual));
             }
         }
@@ -1442,7 +1442,7 @@ namespace FancyZonesUnitTests
             TEST_METHOD (CustomZoneSetsParseSingle)
             {
                 const std::wstring zoneUuid = L"{33A2B101-06E0-437B-A61E-CDBECF502906}";
-                GridLayoutInfo grid(GridLayoutInfo(JSONHelpers::GridLayoutInfo::Full{
+                GridLayoutInfo grid(GridLayoutInfo(FancyZonesDataNS::GridLayoutInfo::Full{
                     .rows = 1,
                     .columns = 3,
                     .rowsPercents = { 10000 },
@@ -1476,7 +1476,7 @@ namespace FancyZonesUnitTests
             {
                 json::JsonObject json;
                 json::JsonArray array;
-                const GridLayoutInfo grid(GridLayoutInfo(JSONHelpers::GridLayoutInfo::Full{
+                const GridLayoutInfo grid(GridLayoutInfo(FancyZonesDataNS::GridLayoutInfo::Full{
                     .rows = 1,
                     .columns = 3,
                     .rowsPercents = { 10000 },
@@ -1548,7 +1548,7 @@ namespace FancyZonesUnitTests
             TEST_METHOD (CustomZoneSetsSerializeSingle)
             {
                 json::JsonArray expected;
-                const GridLayoutInfo grid(GridLayoutInfo(JSONHelpers::GridLayoutInfo::Full{
+                const GridLayoutInfo grid(GridLayoutInfo(FancyZonesDataNS::GridLayoutInfo::Full{
                     .rows = 1,
                     .columns = 3,
                     .rowsPercents = { 10000 },
@@ -1570,7 +1570,7 @@ namespace FancyZonesUnitTests
             {
                 json::JsonObject json;
                 json::JsonArray expected;
-                const GridLayoutInfo grid(GridLayoutInfo(JSONHelpers::GridLayoutInfo::Full{
+                const GridLayoutInfo grid(GridLayoutInfo(FancyZonesDataNS::GridLayoutInfo::Full{
                     .rows = 1,
                     .columns = 3,
                     .rowsPercents = { 10000 },
@@ -1620,7 +1620,7 @@ namespace FancyZonesUnitTests
                 }
 
                 const std::wstring uuid = L"{33A2B101-06E0-437B-A61E-CDBECF502906}";
-                const GridLayoutInfo grid(GridLayoutInfo(JSONHelpers::GridLayoutInfo::Full{
+                const GridLayoutInfo grid(GridLayoutInfo(FancyZonesDataNS::GridLayoutInfo::Full{
                     .rows = 1,
                     .columns = 3,
                     .rowsPercents = { 10000 },
@@ -1675,7 +1675,7 @@ namespace FancyZonesUnitTests
                 json.SetNamedValue(L"devices", devices);
                 data.ParseDeviceInfos(json);
 
-                JSONHelpers::ZoneSetData expectedZoneSetData{
+                FancyZonesDataNS::ZoneSetData expectedZoneSetData{
                     .uuid = L"{33A2B101-06E0-437B-A61E-CDBECF502906}",
                     .type = ZoneSetLayoutType::Focus
                 };
@@ -1700,7 +1700,7 @@ namespace FancyZonesUnitTests
                 json.SetNamedValue(L"devices", devices);
                 data.ParseDeviceInfos(json);
 
-                JSONHelpers::ZoneSetData expectedZoneSetData{
+                FancyZonesDataNS::ZoneSetData expectedZoneSetData{
                     .uuid = L"",
                     .type = ZoneSetLayoutType::Focus
                 };
@@ -1727,7 +1727,7 @@ namespace FancyZonesUnitTests
                 bool parseRes = data.ParseDeviceInfos(json);
                 Assert::IsTrue(parseRes);
 
-                JSONHelpers::ZoneSetData zoneSetData{
+                FancyZonesDataNS::ZoneSetData zoneSetData{
                     .uuid = L"new_uuid",
                     .type = ZoneSetLayoutType::Focus
                 };
@@ -1752,7 +1752,7 @@ namespace FancyZonesUnitTests
                     std::filesystem::remove(jsonPath);
                 }
 
-                const GridLayoutInfo grid(GridLayoutInfo(JSONHelpers::GridLayoutInfo::Full{
+                const GridLayoutInfo grid(GridLayoutInfo(FancyZonesDataNS::GridLayoutInfo::Full{
                     .rows = 1,
                     .columns = 3,
                     .rowsPercents = { 10000 },
